// src/types/tool-names.enum.ts
export enum ToolNames {
  CREATE_TIMEWEB_APP = "create_timeweb_app",
  GET_ALLOWED_PRESETS = "get_allowed_presets",
  ADD_VCS_PROVIDER = "add_vcs_provider",
  GET_VCS_PROVIDERS = "get_vcs_providers",
  GET_VCS_PROVIDER_REPOSITORIES = "get_vcs_provider_repositories",
  GET_VCS_PROVIDER_BY_REPOSITORY_URL = "get_vcs_provider_by_repository_url",
  GET_DEPLOY_SETTINGS = "get_deploy_settings",
}